Refactor stylesheets and add HamburgerButton component
