+++ src/assets/stylesheets/global/_boilerplate.scss
@@ -21,7 +21,6 @@ body {
 
   .app {
     min-height: 100vh;
-    display: flex;
   }
 
   .container {
