+++ src/assets/stylesheets/global/_typography.scss
@@ -20,6 +20,13 @@ a:active {
   color: inherit;
 }
 
+button {
+  background: none;
+  border: none;
+  padding: 0;
+  cursor: pointer;
+}
+
 ul {
   list-style: none;
   padding: 0;
@@ -33,6 +40,5 @@ header {
     left: 0;
     right: 0;
     z-index: 1000;
-    padding: 1rem;
   }
 }
