--- src/components/MobileMenu/index.scss
@@ -1,132 +1,0 @@
-@use '../../assets/stylesheets/util/' as *;
-
-.menu {
-  --bg-menu: var(--text-100);
-  --text: var(--bg-100);
-
-  position: fixed;
-  top: 0;
-  right: -5rem;
-  height: 100vh;
-  width: 75%;
-  box-shadow: -10px 0 10px rgba(0, 0, 0, 0.1);
-  color: var(--text);
-  background-color: rgba(255, 255, 255, 0.2);
-  backdrop-filter: blur(5px);
-  overflow-y: auto;
-  visibility: hidden;
-
-  @include breakpoint-up('md') {
-    display: none;
-  }
-
-  &.open {
-    animation: slide-in 500ms ease-in-out forwards;
-  }
-
-  &.closed {
-    animation: slide-out 500ms ease-in-out forwards;
-  }
-
-  &__button {
-    --dimension: 2.25rem;
-    --font-size: 1.75rem;
-
-    display: none;
-
-    &.remove-button-for-desktop {
-      @include breakpoint-down('sm') {
-        display: block;
-      }
-    }
-
-    &.active {
-      animation: appear 500ms ease-in-out forwards,
-        move-left 500ms ease-in-out forwards;
-    }
-
-    &.inactive {
-      animation: disappear 500ms ease-in-out forwards,
-        move-right 500ms ease-in-out forwards;
-    }
-
-    button {
-      margin-left: auto;
-      background: none;
-      border: none;
-      width: var(--dimension);
-      height: var(--dimension);
-      font-size: var(--font-size);
-      border-radius: 50%;
-      cursor: pointer;
-
-      &:hover {
-        background-color: rgba(255, 255, 255, 0.2);
-      }
-    }
-
-    &--open {
-      --dimension: 3.5rem;
-      --font-size: 2.5rem;
-
-      color: var(--text-200);
-    }
-
-    &--close {
-      translate: 1.5rem 0;
-
-      &.roll-left {
-        animation: roll-left 600ms ease-in-out forwards;
-      }
-
-      &.roll-right {
-        animation: roll-right 700ms ease-in-out forwards;
-      }
-    }
-  }
-
-  &__separator {
-    height: 2px;
-    width: 100%;
-    border-radius: 5rem;
-    background-color: var(--text);
-    opacity: 0.2;
-  }
-
-  &__item {
-    font-size: 1.5rem;
-    margin-top: 0.5rem;
-    position: relative;
-
-    a {
-      width: 100%;
-      border-radius: 0.25rem;
-      transition: background-color 100ms ease-in-out;
-
-      &::before {
-        content: '';
-        position: absolute;
-        top: 0;
-        left: 0;
-        bottom: 0;
-        width: 0.25rem;
-        border-radius: 0.25rem 0 0 0.25rem;
-        background: linear-gradient(
-          45deg,
-          var(--primary-100),
-          var(--primary-300)
-        );
-        opacity: 0;
-        transition: opacity 100ms ease-in-out;
-      }
-
-      &:hover {
-        background-color: rgba(255, 255, 255, 0.2);
-
-        &::before {
-          opacity: 0.85;
-        }
-      }
-    }
-  }
-}
