+++ src/components/Nav/index.jsx
@@ -1,36 +1,118 @@
+import { useRef, useEffect } from 'react';
+import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 
-import Logo from '../globals/Logo';
+import { IoClose as CloseMenuIcon } from 'react-icons/io5';
+import { RiHome3Line as HomeIcon } from 'react-icons/ri';
+import { RiUser3Line as AboutMeIcon } from 'react-icons/ri';
+import { RiFileList2Line as ProjectsIcon } from 'react-icons/ri';
+import { VscTools as CapabilitiesIcon } from 'react-icons/vsc';
+import { RiKakaoTalkLine as ContactIcon } from 'react-icons/ri';
+
 import SocialList from '../globals/SocialList';
+import Logo from '../globals/Logo';
 
 import './index.scss';
 
-import photo from '../../assets/self.JPG';
+const menuItems = [
+  {
+    id: 1,
+    icon: <HomeIcon />,
+    name: 'Home',
+    url: '/',
+  },
+  {
+    id: 2,
+    icon: <AboutMeIcon />,
+    name: 'About Me',
+    url: '/about',
+  },
+  {
+    id: 3,
+    icon: <ProjectsIcon />,
+    name: 'Projects',
+    url: '/projects',
+  },
+  {
+    id: 4,
+    icon: <CapabilitiesIcon />,
+    name: 'Capabilities',
+    url: '/capabilities',
+  },
+  {
+    id: 5,
+    icon: <ContactIcon />,
+    name: 'Contact',
+    url: '/contact',
+  },
+];
 
-const Nav = () => {
+const Nav = ({ handleClick, isMenuOpen }) => {
+  const menuRef = useRef();
+
+  useEffect(() => {
+    const handleClickOutside = (event) => {
+      if (
+        menuRef.current &&
+        !menuRef.current.contains(event.target) &&
+        isMenuOpen
+      ) {
+        handleClick();
+      }
+    };
+
+    document.addEventListener('mousedown', handleClickOutside);
+
+    return () => {
+      document.removeEventListener('mousedown', handleClickOutside);
+    };
+  }, [handleClick, isMenuOpen]);
+
   return (
-    <nav className="nav">
-      <div className="container flex flex-col">
+    <nav className="nav p-3" ref={menuRef}>
+      <div className="nav__button flex flex-jc-sb">
         <Logo />
 
-        <img src={photo} alt="profile" className="nav__profile--photo" />
-
-        <ul className="nav__links">
-          <li className="nav__links--item">
-            <Link to="/">Home</Link>
-          </li>
-          <li className="nav__links--item">
-            <Link to="/page2">Page 2</Link>
-          </li>
-          <li className="nav__links--item">
-            <Link to="/page3">Page 3</Link>
-          </li>
-        </ul>
-
-        <SocialList className="flex flex-col" />
+        <button
+          type="button"
+          className={`nav__close-menu grid grid-pi-c ${
+            isMenuOpen ? 'roll-left' : 'roll-right'
+          }`}
+          onClick={handleClick}
+        >
+          <CloseMenuIcon />
+        </button>
       </div>
+
+      <div className="nav__social py-1">
+        <SocialList className="social-lg flex gap-2" />
+      </div>
+
+      <div className="nav__separator" />
+
+      <ul className="nav__list">
+        {menuItems.map((item) => {
+          return (
+            <li key={item.id} className="nav__item">
+              <Link
+                to={item.url}
+                className="flex flex-ai-c gap-1 p-1"
+                onClick={handleClick}
+              >
+                {item.icon}
+                {item.name}
+              </Link>
+            </li>
+          );
+        })}
+      </ul>
     </nav>
   );
+};
+
+Nav.propTypes = {
+  handleClick: PropTypes.func.isRequired,
+  isMenuOpen: PropTypes.bool.isRequired,
 };
 
 export default Nav;
